pyproject.toml
[project]
name = "customer-onboarding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.2.8",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
]
